services/payments/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
email-validator==2.3.0
pydantic==2.12.5
appwrite==6.0.0
python-dotenv==1.2.1
httpx==0.28.1